+++ tournament_manager2/app/tests/db_utils.py
@@ -1,0 +1,15 @@
+from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+from sqlalchemy.orm import sessionmaker
+from models.base import Base
+
+async def get_db_session(): 
+    engine = create_async_engine('sqlite+aiosqlite:///:memory:', echo=False)
+    async_session = sessionmaker(
+        bind=engine, class_=AsyncSession, expire_on_commit=False
+    )
+
+    # Create tables in the in-memory database
+    async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+
+    return async_session
